app/functions/functions/requirements.txt
firebase_functions~=0.1.0
firebase-admin~=6.2.0
numpy~=1.24.0
Pillow~=10.0.0
tflite-runtime~=2.13.0